feat: setup typescript, prettier and eslint
