+++ tsconfig.json
@@ -1,0 +1,28 @@
+{
+  "$schema": "https://json.schemastore.org/tsconfig",
+  "compilerOptions": {
+    "module": "ES2022",
+    "target": "ES2023",
+    "moduleResolution": "Bundler",
+    "resolveJsonModule": true,
+    "importHelpers": true,
+    "allowJs": true,
+    "esModuleInterop": true,
+    "isolatedModules": true,
+    "skipLibCheck": true,
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "verbatimModuleSyntax": true,
+    "allowSyntheticDefaultImports": true,
+    "removeComments": true,
+    "baseUrl": ".",
+    "paths": {
+      "@/*": ["src/*"]
+    },
+    "types": ["node"],
+    "lib": ["ES2023"]
+  },
+  "exclude": ["node_modules", "front"],
+  "include": ["src/**/*.ts"]
+}
